+++ src/pretalx/settings.py
@@ -288,11 +288,11 @@ USE_L10N = True
 USE_TZ = True
 LANGUAGES = [('en', _('English')), ('de', _('German')), ('fr', _('French'))]
 LANGUAGES_NATURAL_NAMES = [('en', 'English'), ('de', 'Deutsch'), ('fr', 'Français')]
-LANGUAGE_CODE = 'en'
 LANGUAGES_OFFICIAL = {'en', 'de'}
 LOCALE_PATHS = (os.path.join(os.path.dirname(__file__), 'locale'),)
 FORMAT_MODULE_PATH = ['pretalx.common.formats']
-TIME_ZONE = 'UTC'  # TODO
+TIME_ZONE = config.get('locale', 'time_zone')
+LANGUAGE_CODE = config.get('locale', 'language_code')
 
 
 ## AUTHENTICATION SETTINGS